server/pyproject.toml
# AG-UI ADK ミドルウェアのパッケージ定義
[project]
name = "ag_ui_adk"
version = "0.3.2"
authors = [
    { name = "Mark Fogle", email = "mark@contextable.com" }
]
requires-python = ">=3.9, <3.14"
dependencies = [
    "ag-ui-protocol>=0.1.10",  # AG-UI プロトコル
    "aiohttp>=3.12.0",
    "asyncio>=3.4.3",
    "fastapi>=0.115.2",        # Web フレームワーク
    "google-adk>=1.18.0",      # Google ADK
    "google-genai>=1.39.0",    # Gemini API
    "litellm>=1.80.0",         # マルチ LLM ラッパー
    "pydantic>=2.11.7",        # 設定バリデーション
    "pydantic-settings>=2.6.1",
    "pyjson5>=1.6.6",          # JSON5 パーサー
    "uvicorn>=0.35.0",         # ASGI サーバー
]

# ビルドシステム（uv を使用）
[build-system]
requires = ["uv_build>=0.5.0"]
build-backend = "uv_build"

# 開発用依存パッケージ
[dependency-groups]
dev = [
    "black>=25.1.0",           # コードフォーマッタ
    "flake8>=7.3.0",           # リンター
    "isort>=6.0.1",            # import 整列
    "mypy>=1.16.1",            # 型チェック
    "pluggy>=1.6.0",
    "pytest>=8.4.1",           # テストフレームワーク
    "pytest-asyncio>=1.0.0",
    "pytest-cov>=6.2.1",       # カバレッジ
]
